+++ projects/jsf-validation/src/lib/schema-validation.service.spec.ts
@@ -1,4 +1,5 @@
-import { SchemaValidationService } from './schema-validation.service';
+import { AdditionalPropertiesParams } from 'ajv';
+import { JSFErrorObject, SchemaValidationService } from './schema-validation.service';
 
 describe('SchemaValidationService', () => {
   let basicSchema: any;
@@ -6,7 +7,7 @@ describe('SchemaValidationService', () => {
   beforeEach(() => {
     basicSchema = {
       type: 'object',
-      required: ['checkboxInput', 'numberInput'],
+      required: ['checkboxInput', 'numberInput', 'arrayInput', 'defaultedRequiredValue'],
       properties: {
         checkboxInput: {
           type: 'boolean',
@@ -24,69 +25,263 @@ describe('SchemaValidationService', () => {
           minLength: 5,
           maxLength: 10
         },
-        hiddenInput: {
+        defaultedRequiredValue: {
           type: 'string',
-          name: 'Hidden field should not be validated',
-          isHidden: true,
-          maxLength: 3,
-          default: 'too long'
+          name: 'Default Text Input',
+          default: 'should not cause error when value is omitted'
         },
-        readonlyInput: {
-          type: 'string',
-          name: 'Readonly field should not be validated',
+        arrayInput: {
+          type: 'array',
+          name: 'Array field',
+          items: {
+            properties: {
+              textItem: {
+                type: 'string',
+                name: 'Text Item',
+                minLength: 5,
+                maxLength: 10
+              },
+              requiredItem: {
+                type: 'string',
+                name: 'Required Item'
+              },
+              requiredItemWithDefault: {
+                type: 'string',
+                name: 'Default Required Item',
+                default: 'should not cause error when value is omitted'
+              }
+            },
+            required: ['requiredItem', 'requiredItemWithDefault']
+          }
+        },
+        templateInput: {
+          type: 'object',
+          name: 'Template Test',
+          properties: {
+            templateDisplay: {
+              name: 'Template 1',
+              type: 'template',
+              isReadOnly: true,
+              templateName: 'testTemplate1',
+              targetPaths: [
+                'templateInput.templateValue',
+                'templateInput.templateVisibleValue'
+              ]
+            },
+            templateValue: {
+              name: 'Hidden template value',
+              type: 'string',
+              isHidden: true
+            },
+            templateVisibleValue: {
+              name: 'Visible template value',
+              type: 'string',
+              isHidden: false
+            }
+          },
+          additionalProperties: false
+        },
+        buttonInput: {
+          name: 'button 1',
+          type: 'button',
           isReadOnly: true,
-          maxLength: 3,
-          default: 'too long'
+          targetPaths: [
+            'buttonValue'
+          ]
+        },
+        buttonValue: {
+          name: 'Hidden value',
+          type: 'string',
+          isHidden: true
         }
       }
     };
   });
 
   describe('validate()', () => {
-    it('should return an array of errors when missing required properties', () => {
-      const result = SchemaValidationService.validate(basicSchema, { textInput: 'abcde' });
-      expect(result[0].errorObject.message).toContain('checkboxInput');
-      expect(result[1].errorObject.message).toContain('numberInput');
+    it('back end validation should validate uri format, and at different depths', () => {
+      let err: JSFErrorObject[] = [];
+
+      let uriSchema: any = {
+        type: 'object',
+        properties: {
+          inbound: {
+            name: 'inbound',
+            type: 'object',
+            properties: {
+              url: {
+                type: 'string',
+                name: 'my uri string',
+                format: 'uri'
+              }
+            }
+          }
+        }
+      };
+      err = SchemaValidationService.validate(uriSchema, {inbound: {url: 'https://t'}});
+      expect(err[0].errorObject.schemaPath).toBe('properties.inbound.properties.url.format');
+      err = SchemaValidationService.validate(uriSchema, {inbound: {url: 'https://test.com'}});
+      expect(err).toBe(null);
+
+      uriSchema = {
+        type: 'object',
+        properties: {
+          url: {
+            type: 'string',
+            name: 'my uri string',
+            format: 'uri'
+          }
+        }
+      };
+      err = SchemaValidationService.validate(uriSchema, {url: 'https://t'});
+      expect(err[0].errorObject.schemaPath).toBe('properties.url.format');
+      err = SchemaValidationService.validate(uriSchema, {url: 'https://test.com'});
+      expect(err).toBe(null);
+
+      const urls: [string, boolean][] = [
+        ['test.com', false],
+        ['https://test.com', true],
+        ['https://test.com:4200', true],
+        ['ftp://test.com', true],
+        ['http://test.com?query=testquery', true],
+        ['http://test-hyphen.com', true],
+        ['http://test.com', true],
+        ['http://test.co', true],
+        ['test.com/🤔', false],
+        ['http://test.com/🤔', false],
+        ['http://test.com/', true],
+        ['test.com/text\u0002.com', false],
+        ['http://test.com/text\u0002.com', false],
+        ['http://test.com/textu0002.com', true],
+        ['http://test.c', false],
+        ['test', false],
+        ['http://test .com', false],
+        ['https//test.com', false],
+        ['http://i-comms.truecommerce.net:42000AS2/F30510A3CH', false],
+        ['http://i-comms.truecommerce.net:42000/AS2/F30510A3CH', true]
+      ];
+
+      for (const testSet of urls) {
+        const url = testSet[0];
+        const expectedValidity = testSet[1];
+        const validationError = SchemaValidationService.validate(uriSchema, {url: url});
+        expect(validationError == null).toBe(expectedValidity, `url: ${url} expectedValidity: ${expectedValidity}`);
+      }
     });
 
-    it('should return an array of errors when missing one properties', () => {
-      const result = SchemaValidationService.validate(basicSchema, { textInput: 'abcde', checkboxInput: true});
+    it('should return an array of errors when missing required properties', () => {
+      const result = SchemaValidationService.validate(basicSchema, {textInput: 'abcde'});
+      expect(result[0].errorObject.message).toContain('checkboxInput');
+      expect(result[1].errorObject.message).toContain('numberInput');
+      expect(result[2].errorObject.message).toContain('arrayInput');
+    });
+
+    it('should return an array of errors when missing one property', () => {
+      const result = SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          checkboxInput: true,
+          arrayInput: [{textItem: 'abcde', requiredItem: 'abcde'}]
+        }
+      );
       expect(result[0].errorObject.message).toContain('numberInput');
     });
 
     it('should return an null when values are valid', () => {
-      const result = SchemaValidationService.validate(basicSchema, { textInput: 'abcde', numberInput: 0, checkboxInput: true});
+      const result = SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          numberInput: 0,
+          checkboxInput: true,
+          arrayInput: [{textItem: 'abcde', requiredItem: 'abcde'}]
+        }
+      );
       expect(result).toBeNull();
     });
 
     it('should return an array of errors when invalid properties', () => {
-      const result = SchemaValidationService.validate(basicSchema, { textInput: 'abc', numberInput: 13, checkboxInput: false });
+      const result = SchemaValidationService.validate(basicSchema, {textInput: 'abc', numberInput: 13, checkboxInput: false});
       expect(result[0].errorObject.message).toContain('should be <= 12');
       expect(result[1].errorObject.message).toContain('should NOT be shorter than 5 characters');
+      expect(result[2].errorObject.message).toContain("should have required property 'arrayInput'");
     });
+
+    it('should return an error when array value is invalid', () => {
+      const result = SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          numberInput: 0,
+          checkboxInput: true,
+          arrayInput: [{textItem: 'abc', requiredItem: 'abcde'}]
+        }
+      );
+      expect(result[0].errorObject.message).toContain('should NOT be shorter than 5 characters');
+    });
+
+    it('should throw an error when template type values are set', () => {
+      const result = SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          checkboxInput: true,
+          arrayInput: [{textItem: 'abcde', requiredItem: 'abcde'}],
+          numberInput: 1,
+          templateInput: {templateDisplay: '1', templateVisibleValue: '2', templateValue: '3'}
+        }
+      );
+      expect((result[0].errorObject.params as AdditionalPropertiesParams).additionalProperty).toContain('templateDisplay');
+    });
+
   });
 
   describe('prettyPrintErrors()', () => {
     it('should format an array of errors when missing required properties', () => {
-      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, { textInput: 'abcde' }));
+      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, {textInput: 'abcde'}));
       expect(result).toContain('Checkbox Input is required.');
       expect(result).toContain('Number Input is required.');
+      expect(result).toContain('Array field is required.');
     });
 
-    it('should format an array of errors when missing one properties', () => {
-      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, { textInput: 'abcde', checkboxInput: true}));
+    it('should format an array of errors when missing properties', () => {
+      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, {textInput: 'abcde', checkboxInput: true}));
       expect(result).toContain('Number Input is required.');
+      expect(result).toContain('Array field is required.');
     });
 
     it('should not format errors when values are valid', () => {
-      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, { textInput: 'abcde', numberInput: 0, checkboxInput: true}));
+      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          numberInput: 0,
+          checkboxInput: true,
+          arrayInput: [{textItem: 'abcde', requiredItem: 'abcde'}]
+        }
+      ));
       expect(result).toEqual('');
     });
 
     it('should format an array of errors when invalid properties', () => {
-      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema, { textInput: 'abc', numberInput: 13, checkboxInput: false }));
+      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abc',
+          numberInput: 13,
+          checkboxInput: false
+        }
+      ));
       expect(result).toContain('Text Input should NOT be shorter than 5 characters.');
       expect(result).toContain('Number Input should be <= 12.');
+      expect(result).toContain('Array field is required.');
+    });
+
+    it('should format an array of errors when invalid array item', () => {
+      const result = SchemaValidationService.prettyPrintErrors(SchemaValidationService.validate(basicSchema,
+        {
+          textInput: 'abcde',
+          numberInput: 10,
+          checkboxInput: false,
+          arrayInput: [{textItem: 'abcde'}]
+        }
+      ));
+      expect(result).toContain('Required Item is required.');
     });
   });
 });